Fix nulls bitmap size when number of columns is more than 64 (YQL-17952) (#2422)

+++ ydb/library/yql/minikql/comp_nodes/mkql_grace_join.cpp
@@ -272,7 +272,7 @@ void TGraceJoinPacker::Pack()  {
         case NUdf::EDataSlot::Interval:
             WriteUnaligned<i64>(buffPtr, value.Get<i64>()); break;
         case NUdf::EDataSlot::Date32:
-            WriteUnaligned<i64>(buffPtr, value.Get<i32>()); break;
+            WriteUnaligned<i32>(buffPtr, value.Get<i32>()); break;
         case NUdf::EDataSlot::Datetime64:
             WriteUnaligned<i64>(buffPtr, value.Get<i64>()); break;
         case NUdf::EDataSlot::Timestamp64:
